requirements.txt
asgiref==3.4.1
attrs==21.4.0
black==22.1.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==35.0.0
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
h11==0.9.0
httpcore==0.11.1
httpx==0.15.5
idna==3.3
jsonschema==3.2.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.10.4
PyJWT==2.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2021.3
redis==4.1.4
requests==2.27.1
requests-file==1.5.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.14.2
supertokens-python==0.5.2
tldextract==3.1.0
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.14.0
